Add logic for handling a tmpdir that wasn't created when a path is provided to -tmp
This can happen when a tmpdir isn't necessary to create because the necessary files are already present in the -db path.

+++ confindr_src/confindr.py
@@ -118,7 +118,8 @@ def confindr(args):
             if args.keep_files is False:
                 shutil.rmtree(os.path.join(args.output_name, sample_name))
     if args.keep_files is False and args.tmp is not None:
-        shutil.rmtree(args.tmp)
+        if os.path.isdir(args.tmp):
+            shutil.rmtree(args.tmp)
     logging.info('Contamination detection complete!')
